Attack Mode
The robot will now push the puck back if it is stationary and within reach.  Also, if the puck is moving within the range which the robot can reach at a slow speed, the code predicts an intercept point an equidistance away from the current position as the previous saved position.  This code does not currently take into account puck bounce.