attack mode, striker positioning
Checks coordinated of the predicted intercept location for the moving puck in attack mode.  If it's off the table, striker does not bother moving.

Noticed when puck disappeared occasionally the motors continued to drive the striker.  Added a condition that if no fuschia contours are found, send the coordinates for rest position.  To account for this, the try-except block was moved out.